pie-docs-backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
psycopg2-binary==2.9.10
asyncpg==0.30.0
sqlalchemy>=1.4,<2.0.36

# Vector embeddings
openai==1.54.4
sentence-transformers==3.2.1
tiktoken==0.8.0

# RAG and document processing
langchain==0.3.7
langchain-community==0.3.5
langchain-openai==0.2.5

# PDF to Image conversion for AI classification
pdf2image==1.17.0
pillow==11.3.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
numpy==1.26.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
